1, s3cipid_t is upgraded to int16 because we need that, I already check that there are no magic code using 8-bit s3cipid_t 2, Refactor the ep code and put a lot of stuffs into fe.c (should be renamed to something else. 3, Check-in codes of wave2feat and cepview. (cepview will not dump core but Evandro will kill me) 4, Make the same command line frontends for decode, align, dag, astar, allphone, decode_anytopo and ep . Allow the use a file to configure the application. 5, Make changes in test such that test-allphone becomes a repeatability test. 6, cepview, wave2feat and decode_anytopo will not be installed in 3.5 RCIII (Known bugs after this commit) 1, decode_anytopo has strange bugs in some situations that it cannot find the end of the lattice. This is urgent. 2, default argument file's mechanism is not yet supported, we need to fix it. 3, the bug discovered by SonicFoundry is still not fixed.
git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk/sphinx3@4129 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ src/libs3decoder/flat_fwd.h
@@ -109,7 +109,7 @@ typedef struct whmm_s {
     s3pid_t   *pid;		/* Triphone id: 1 per state if 1st phone in word,
 				   otherwise single pid for entire phone */
     int32      bestscore;	/* Best among this whmm.score[] in current frame */
-    int8       pos;		/* Word pronunciation position index */
+    int16       pos;		/* Word pronunciation position index */
     s3cipid_t  rc;		/* Right context position (only for last phone in word);
 				   index into rcpid[][].pid or lrcpid[][].pid */
     int32      active;		/* Whether active in current frame */
